ui - refactor prompt editor
