+++ ui/litellm-dashboard/src/components/prompts/prompt_editor_view/PromptMessagesCard.tsx
@@ -1,0 +1,78 @@
+import React from "react";
+import { Card, Text } from "@tremor/react";
+import { Select } from "antd";
+import { PlusIcon, TrashIcon } from "lucide-react";
+import VariableTextArea from "../variable_textarea";
+import { Message } from "./types";
+
+const { Option } = Select;
+
+interface PromptMessagesCardProps {
+  messages: Message[];
+  onAddMessage: () => void;
+  onUpdateMessage: (index: number, field: "role" | "content", value: string) => void;
+  onRemoveMessage: (index: number) => void;
+}
+
+const PromptMessagesCard: React.FC<PromptMessagesCardProps> = ({
+  messages,
+  onAddMessage,
+  onUpdateMessage,
+  onRemoveMessage,
+}) => {
+  return (
+    <Card className="p-3">
+      <div className="mb-2">
+        <Text className="text-sm font-medium">Prompt messages</Text>
+        <Text className="text-gray-500 text-xs mt-1">
+          Use <code className="bg-gray-100 px-1 rounded text-xs">{'{{variable}}'}</code> syntax for template variables
+        </Text>
+      </div>
+      <div className="space-y-2">
+        {messages.map((message, index) => (
+          <div key={index} className="border border-gray-300 rounded overflow-hidden bg-white">
+            <div className="bg-gray-50 px-2 py-1.5 border-b border-gray-300 flex items-center justify-between">
+              <Select
+                value={message.role}
+                onChange={(value) => onUpdateMessage(index, "role", value)}
+                style={{ width: 100 }}
+                size="small"
+                bordered={false}
+              >
+                <Option value="user">User</Option>
+                <Option value="assistant">Assistant</Option>
+                <Option value="system">System</Option>
+              </Select>
+              {messages.length > 1 && (
+                <button
+                  onClick={() => onRemoveMessage(index)}
+                  className="text-gray-400 hover:text-red-500"
+                >
+                  <TrashIcon size={14} />
+                </button>
+              )}
+            </div>
+            <div className="p-2">
+              <VariableTextArea
+                value={message.content}
+                onChange={(value) => onUpdateMessage(index, "content", value)}
+                rows={3}
+                placeholder="Enter prompt content..."
+              />
+            </div>
+          </div>
+        ))}
+      </div>
+      <button
+        onClick={onAddMessage}
+        className="mt-2 text-xs text-blue-600 hover:text-blue-700 flex items-center"
+      >
+        <PlusIcon size={14} className="mr-1" />
+        Add message
+      </button>
+    </Card>
+  );
+};
+
+export default PromptMessagesCard;
+
